+++ app/src/main/java/org/android/go/sopt/home/GalleryFragment.kt
@@ -5,12 +5,9 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import org.android.go.sopt.R
-import org.android.go.sopt.adapter.GalleryAdapter
+import org.android.go.sopt.home.adapter.GalleryAdapter
 import org.android.go.sopt.databinding.FragmentGalleryBinding
 import org.android.go.sopt.home.viewmodel.GalleryViewModel
-import org.android.go.sopt.home.viewmodel.HomeViewModel
-import org.android.go.sopt.sampledata.Repo
 
 class GalleryFragment: Fragment() {
     val itemList : List<Int> = GalleryViewModel().getImgList() //galleryViewModel img 보관
