+++ app/src/main/java/org/android/go/sopt/home/adapter/TitleAdapter.kt
