+++ app/src/main/res/layout/activity_fragmentmanage.xml
@@ -5,7 +5,7 @@
     android:padding="10dp"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".FragmentManageActivity">
+    tools:context=".home.FragmentManageActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fcv_main"
