+++ app/src/main/res/layout/activity_sign_up.xml
@@ -10,7 +10,7 @@
         android:layout_height="match_parent"
         android:paddingVertical="30dp"
         android:paddingHorizontal="30dp"
-        tools:context=".SignUpActivity">
+        tools:context=".signUp.SignUpActivity">
 
         <TextView
             android:id="@+id/textview_signup_title"
